Initial commit
Basic working code besides the terminal interface. Not really here for public consumption but rather to tracking code changes over time. This code is designed to work with a bespoke antenna mount I designed. The idea is to create an lighweight azimuth/elevation mount that can accept different instruments, besides just parabolic dish antennas. Using TLE data the mount should be able to track earth-orbiting objects. Future ideas are to allow tracking of any celestial objects with other instruments such as a high-power laser. The idea there being you could tell the computer to point to a start or planet and it would move the laser onto it and flash it a few times.

At any rate this code is mostly complete for what it is: basic movement control over both axes. It uses a closed-loop movement control to track the current position vs the commanded position and constantly keeps itself pointed with an accuracy of 0.15 degrees on the azimuth axis and 0.18 degrees on the elevation axis.

The initial hope with this project is to collect HRPT data from polar-orbiting earth observation/weather satellites - which requires a high gain dish and accurate tracking. Most amateurs do the tracking by hand by I'm lazy so I'm hoping this design will do it for me. We'll see...