Basic sat finder functionality
Started work on the sat finder class. For now it can list the passes for any satellite in its TLE file (which it will keep updated) over the next 24 hours. It has mostly the same data WXtoIMG has, including its max elevation, east/west, duration, etc. For now the code can be told to initiate a track and will move to its starting position but won't actually track. That part comes next, should be pretty easy. Just a countdown to the time and then constantly updating the azimuth/elevation commanded angles. Will have to track the elevation and stop tracking when it gets back down to 0 again.