Big mistake on azimuth direction
I have no idea how I failed to notice that the azimuth was rotating counterclockwise through its positive rotation but I finally DID notice it after I set up, leveled, and aligned the antenna for a test. I needed to change a LOT to get it to work but in the end I think all I really needed to do was swap the azimuth anglesigns but I don't know for sure now. And it works so it doesn't really matter now.

Another issue I just stumbled upon, is my old friend the 0/360 degree boundary. Some passes will transit through zero degrees and the direction code is too dumb to handle that. It tries to spin around the opposite way to get to the 300s instead of knowing that it can get to the 300s by just going through 0. It just needs to be made away that it can simply go through zero to 359.

The other issue arising from moving from 0 to 360 is that we could move past the azimuth endstop. If we end up on the other side we would go home the wrong direction, winding up further. We will have to track the direction and total number of steps moved and just "backtrack" instead of a simple move_to_angle(0).

Getting there, two steps forward and one step back.