Smart azimuth movement
One of the issues I found during the last test was that when the commanded angle transited through 0/359 the move motor function would try to spin around the wrong way to get to that point, instead of going through the 0/359 line straight to the target angle. Updated move_to_commanded_angle() function now figures out if moving the opposite way is actually faster. It also corrects the encoder's current angle and the commanded angle to be within 0-359 and be positive. The big issue now is that its very easy to end up in a situation where we progressively wind the mast up more and more until the wires get pulled. We need to track whether we pass the 0/359 line and if we do we have to unwind PAST that before we can do a normal move_to_angle(0). We really need the multi-turn endstop design for the azimuth axis, that way we can positively track our location with ease.

Also fixed a small bug in the terminal interface where if you didnt enter any valid arguments to the 'go' command it would crash. And the azimuth homing function now uses raw movement commands instead of move_to_angle() to speed up calibration a bit.