Improved watch command
The old watch command was great at tracking the antenna position but the only way to stop the watch mode was the ctrl+c. This improvement tracks the changes on the elevation and azimuth axis and when they stop changing the watch command ends and returns to the prompt. How long it waits after no more changes can be tuned, currently it waits 5 loop iterations and each loop iteration has a sleep delay of 0.1 seconds, so maybe half a second wait.

Also fixed a possible bug that would maybe cause the move_to_commanded_angle() function to never end but only for opposite moves that are ended before reaching the 0/359 line? I dunno, it never caused problems but I fixed it anyway incase it could.