Mast unwind
Added function that unwinds the mast back to its home position, regardless of what side of the endstop its on currently. It uses the incremental encoder to know its true position since the start of the program. Now on program exit it will unwind the mast to be sure we don't end up progressively winding up over time and reboots.