Updates to terminal interface and satfinder
Covered the entire terminal interface loop with a try/except that keeps the loop going if the user ctrl-c's anything. Previously it would exit the app if you were tracking a sat or something deeper in the loop when you ctrl-c'd.

Also realized I wasnt using the correct data with the satfinder output. WXtoIMG gives the longitude of the pass but I was incorrectly showing the azimuth instead. Now correctly shows the longitude with east or west information. Finally, I made the error messages for the passlist (and related commands) to inform when the sat isnt listed vs when its geostationary (which isnt support by pyorbital).