Satfinder improvements
Updated the satfinder class so that it could handle incorrect names in a smarter way. Now we check against our satellite name list and if anything matches close enough we just roll with it. Otherwise we error out like before but also show the user the closest matching satellite name.

Added the pass direction (north or south) to the passlist output. I think now the passlist shows all the same information that WXtoIMG does. Thinking about spinning that off into its own app incase other people wanted it.

Covered the track command's code in its own try/except statement so that I could use a finally: section to unwind the mast no matter what happens, failure or success.

Tried to fix an issue where the status output of the antenna position would cause newlines to be printed instead of clearing and printing over the same line. Turns out if you clear more characters than the current line width it causes a newline to print. Now we check the length of the current print message, add 10 to that, and blank out that many chars. Seems to have fixed it for now.

I keep forgetting to update the version number.