Small operational update
All the testing was done in a small room so I never had the dish attached. With the dish attached by PWM constants are all wrong for the elevation axis. It cant move itself off horizontal at the lowest PWM duty cycle. It barely started to move when ~3 degree away using the old constants so everything got shuffled down. It doesn't seem to be causing any issues with overrunning the target (and even if it did the closed-loop control would bring it back). There are some MAJOR issues with the PWM constants on the azimuth axis when the satellite is passing nearly directly overhead. At the max elevation the aximuth axis can shift like 5+ degrees in a single moment and that makes for really jarring moves. It didnt lose signal but it looks god-awful and I would like to fix it somehow. Need PROPER PID control, all parts of it.

Also just discovered that I can use \t escape chars in my print statements to align everything so I'm definitely doing that next.