Filter display passes
Previously the only way to filter low elevation passes was to set the start elevation to that number, but then you can't track from the horizon. Now there's two separate variables that control both. Old variable still effects the start elevation of a tracking pass, new one limits passes with low max elevation from appearing in the passlist.

Next idea up to do is a power-failure/resume system. If the power ever goes out or the control program crashes, and I'm turned backwards past 0, it will try to find home by winding up a full turn, possibly damaging wires. Basically I have to go down there and manually disconnect the motor and turn it back past the sensor.

Instead I'm imaging a system where the current angle and (more important) tick number are printed out to a file on disk. When the program changes the azimuth or elevation at all, it writes the new location data to file. On program exit it will always unwind and then delete the file. If on startup we find the file there we know we didn't exit cleanly. We can then just "assume" the data in the file is actually still our current position. There could be some minor loss of accuracy if the mast moves at all before restarting the control program - but at least we can recover and unwind without issue. Then we just rehome and we're back in business. If we're pointed at GOES when we crash, and we come back still with perfect signal quality, we can probably skip rehoming.

It will also allow for restarting the program for updates. Currently the only way to update the program is to stop and restart it, requiring a rehome of the entire system - super annoying. And thats the primary reason this code push is completely untested, might be buggy and crash things. I don't like to rehome the system when I'm not there to physically watch it.

Anyway that idea is next.