Better filtering
Filtering worked fine, but there was a problem with the start times and total duration when displaying the passlists. Using the display filter limit with the get_next_passes() function was returning start and end times at that display filter limit instead of at the tracking limit which changed the total pass duration a lot. Now displays the total pass duration based on the tracking limits (not the time from horizon to horizon).