Load GPS coords from file
Instead of loading the GPS coordinates from the python file, it can now read them from a text file in the same dir as the python file. In the future if you needed to update the coordinates without restarting this would allow it. Currently only checks at start for the gpscoords.txt file. Format of the file is the same as the 3 lines in the python file, just no comments.

Also forgot to mention last commit that I added a new command "updatetle" to allow updating the TLE weather.txt file without restarting.