Small bugfix
Didnt test the code before committing, a couple things got by. The error message about not having gpscoords.txt file was always firing and the regex to match the GPS coords in the file couldn't handle negatives. Simple fix for both. Also forgot to update the print statement in passlist() to show the timeframe of the predicted passes (was set to 24h before, now shows the actual timeframe). Finally, I added a new command, "gpscoords", to help verify that the coords are correctly loaded from the file. Might not be a bad idea to have a command to manually change those coords at runtime.