Position recovery
Most important new addition in this commit is the ability to load position data from a recovery file, in case we lose power or crash or need to restart without homing. Its possible to lose a position increment or two on the azimuth axis when recovering. So final azimuth angle could be off by +- 2*ANGLE_TICK. The program isn't currently saving recovery data though, that's next. Its basically just going to save its current position state to a json file every 5 seconds or something. Should be fine.

A bunch of small stuff in this one too. The passlist command is now more verbose about the reason for not finding any sats by mentioning the max elevation limit. The updatetle command now also updates the satellite name list, although its unlikely to change. Finally, changed the homing process's move_until_stop() commands (on both axis) to check its position and endstops 10x faster. I've noticed the tick count change when homing the elevation axis. Sometimes its like 1012, sometimes its 1013. Not sure if its variation in the endstop switches or just poor timing in the loop. Its only running for the homing processes so I don't mind the increase CPU usage for a minute or so just to be sure its not timing. Had to update the print statement for the azimuth axis' move_until_stop() command to account for the fast print rates. If you print too fast it sometimes causes issues with incomplete printing and you can't read the end of the line. The additional code just slows it down with respect to the loop speed.

Also had an idea to implement absolute tick limits so that we can't wind the mast up too far and damage things. It could also allow us to determine if we need to unwind before moving. If we move in one direction less than 180 degrees multiple times, we will wind the axis up and do damage. Looking at the tick count we can decide to just unwind first before further movements. Something to think about.

Untested again so prolly another bugfix coming after this lol.