Rearrange a bit and elevation axis now parks on quit
An end result of the "Just get it to work" mantra when it comes to coding is a huge mess of code thats not well laid out. I mean it all works, but its not fun to look at. Trying to sort out the code a bit to help. Should also fix the terminal interface code so you can actually find the commands in the code. Might even be able to get rid of the move_to_angle() functions and use move_to_commanded_angle() instead. Its only used during homing for the elevation axis and the azimuth only uses it once after unwinding, and just to go to 0.

Also somehow forgot to tell the elevation axis to park itself before quitting.