Movement bug when less less than commanded angle
Looks like in the code section that deals with movements smaller than our ANGLE_TICK, we were comparing an angle with sign information. This seems like it could lead to a situation where if the absolute value of anglediff is larger than ANGLE_TICK but is also a negative value, it would hit the stop_motor() function early. Switched to comparing against the absolute value should eliminate that possibility.

Also adjusted the critical angle again, seemed like it was going out of sync between 15 and 20 degrees, but only by maybe half a degree.

Finally, added the beginnings of a function to manually adjust the azimuth offset. Only testing things for now, trying to figure out the best way to accomplish it.