Didnt help
Still has the issue that it wont move to the commanded angle between 0.4 degrees and angle tick away, but it only seems to happen when descending through the angles. I thought it was because anglediff was negative but the anglesign specifically fixes that negative, so I have no idea why there would be any difference going positive or negative. If somehow the anglediff WAS negative though it would always be at the slowest movement speed which might not move if the mast is tight.

Whats needed to debug this further, besides running in debug mode and tracing every if statement, is to have an internal variable thats true when moving and set false whenever stop_motor() is called. That way I can see if its a stop_motor() call thats ending movement early or if its the duty cycle being too low. Currently outputting that new variable in the debug console command (just letter "d").

Also further fleshed out the setoffset console command but ran into a slight complication. Currently the tick number is used to unwind the mast and if we moved through 0 to get to our current position the tick number is negative instead of positive. When we change the current angle, we have to update the correct tick number depending on whether we are negative or not. I think the best plan is to take the difference in angle, convert that to ticks, and add that offset to the current tick number - instead of trying to determine the tick number from the actual angle and ANGLE_TICK. Just do a relative change. Still thinking that one over.