Offset update
Went over the offset command and pared it down a bit to whats really needed. Elevation axis offsets are dumb and can't be adjusted like that anyway. You change the azimuth and suddenly the elevation offset needed is different. So removed all the elevation code. Also ditched the idea of changing the tick number, we need that to unwind safely anyway. Since ticks arent used for pointing manually or tracking sats, we don't need to update it for what we are trying to accomplish.

A new issue cropped up though, and thats parking. The mount parks both axis before quitting, and it just uses the move_to_angle() command. If the offset was too big it could lead to a situation where it parks on the other side of the hall effect sensor's magnet which is big league bad news. Now we track the offset difference from the current angle and the input offset angle and add that difference back when we park. It should track multiple offset changes correctly, but probably safest to just offset once.