Power failure recovery file loading
Never tested the check_for_recovery_data() function and when I did it had some issues. It couldn't handle json files that spanned multiple lines. Fixed by just loading the file itself instead of reading its lines and loading those. Changed the way the needed keys are checked against the actual keys. Before it was expecting the lists to be exactly the same, down to the order. Now it checks each key in the needed keys against the actual key list and if they are all present it continues. This is tolerant of extra keys and out of order keys. Also changed the way the data was presented to the user for checking over. Previously it was the raw dict type being printed, now its a nice list where the second column is lined up.