Formatting update
Switched to tab separators for some prints to hopefully keep the numbers from jumping around. Added an ETA estimate while waiting and added a percentage readout while tracking a pass that shows where you are in the pass duration. Added checks to the time string creation function so that it only has a plural 's' when necessary.