Small bugfixes
Changed file loading sections to instead use the full absolute file path instead of just the file name. I start the app from symlinks and in different folders, so it was causing issues. Now it needs the path to whereever its running in the RUNFOLDER variable, otherwise it breaks. Could probably fall back to something sane if it fails but for now its fine. Removed some of the tab chars in the print statements as it was causing the print to move around a lot and didn't really buffer the text like I thought it would. Also fixed an issue with the percentage readout when tracking a satellite. I made two mistakes, one was counting the wrong way (going from 100% to 0%) and the other was rounding off a float value to an int BEFORE I multiplied it by 100, so it was always just 100 or 0, and it was always backwards.