Further beating a dead horse
Yeah this is probably the 10th time Ive thought "I should really implement actual PID so I dont have to deal with this anymore" but then I find working constants and everything is gravy for a while. Added in a elevation check for over 80 degrees to lower the PWM constants to try and stop the oscillations between 85-95 when it swings wildly back and forth. Fun times.