New movement test command
Guessing one PWM constant at a time is stupid, so Im going to become more refined and just guess a lot faster. To that end I've added a command that should allow manual control over each axis, allowing you to enter any PWM val and see if it produces movement or not. I should hopefully be able to dial in the needed PWM vals to the minimum across the whole range. At the moment the elevation axis PWM vals are too high and are causing bursts of interference on my receiving radio whenever it moves even a tiny amount so those constants are way too high. I think this should be the end of it, no more critical angle bullcrap either, if I get enough values figured out I can just find a function that fits the values and use it (hopefully, thats only half "handwavy maths will solve all problems").

Also changed some input angle range checks to be inclusive of endpoints, and made the elevation checks not allow values below 0, which would be dumb and probably cause a little damage. What I don't get atm is why the older version allowed me to go to el180 before when the max limit was 180. It didn't use an inclusive check on the max angle so it should have given me crap for entering el180 but it never did. Or maybe I never tested it? I can't remember exactly now. Will test it out when I get physical access again (that way I can unplug if the new code goes spin-cycle on me).