Crash safety
I went to add in the endstop check and found it was already there. I didn't understand why that endstop check didn't work at first, but I think that exception crashed the thread and it was no longer running its move_to_commanded_angle() loop (with the motors were stuck on). The exception that appeared in the console started "Exception in thread Thread-2:" and the stack trace was only code from that class. I don't think the terminal_interface() loop had any error. Thats why the console kept working but the elevation axis didn't respond at all to my inputs or its endstops. I added an exception catch-all to the main terminal_interface() loop even though that wouldn't have helped in this instance. For now I've named the threads to hopefully get slightly better error messages. The specific crash was from the unclamped PWM vals, so that shouldn't happen again. Also had the wrong variable name for the angle, but never was a problem because of the damage stopping me from testing.