Position recovery code v1
Added first iteration of the position recovery code. This code allows the antenna position to be recovered after a power failure, so you don't have to manually unwind and rehome safely. Currently just saves the data when the motors are ordered to stop.

Also added a position check to the beginning of the move_to_commanded_angle loop functions to see if that reduces the idle CPU usage.

Code is untested but looks free of bugs lol.