Fixed wrong GPS altitude and PWM constants
I recently discovered a discrepancy in the reported start times and durations of passes between this control program and my stand-alone satpasslist.py. After going through the code I found that on the control program I wasn't accounting for the fact that the pyorbital library wants the altitude in kilometers, but I use meters. Added in a divide by 1000 to fix it.

Also recently added another pair of reflectors to my antenna, rotated 90 degrees from the first pair, nearly doubling my signal level. But the increased the weight of the antenna made it impossible to move when the difference was less than a degree or two, so we need stronger PWM constants. I didn't really fine tune these well, so they could be over whats needed.

Also added the GPS coordinates to the debug output since thats where I keep looking for it. Totally forgot I already had a gpscoords command lol. Now we have two!